Fix bug on gsheets upload. Refactor chain metadata utils
